requirements.txt
matplotlib==3.8.2
networkx==3.2.1
tqdm==4.66.4
PyYAML==6.0.2
numpy==2.2.6
scikit-learn==1.5.0
scipy==1.13.1
